requirements.txt
streamlit
pandas
langchain>=0.3
langchain-anthropic>=0.1
langchain-groq>=0.1
python-dotenv
tabulate 